+++ appTools/__init__.py
@@ -3,7 +3,7 @@ from appTools.ToolCalculators import ToolCalculator
 from appTools.ToolCalibration import ToolCalibration
 
 from appTools.ToolDblSided import DblSidedTool
-from appTools.ToolExtractDrills import ToolExtractDrills
+from appTools.ToolExtract import ToolExtract
 from appTools.ToolAlignObjects import AlignObjects
 
 from appTools.ToolFilm import Film